refactor: reorganize folder structure and rename files for better clarity
