+++ composeApp/src/desktopTest/kotlin/domain/usecase/GetMovieDetailUseCaseTest.kt
@@ -1,3 +1,5 @@
+package domain.usecase
+
 import edu.dyds.movies.domain.usecase.GetMovieDetailsUseCaseImpl
 import edu.dyds.movies.domain.entity.Movie
 import edu.dyds.movies.domain.repository.MoviesRepository
