+++ composeApp/src/desktopTest/kotlin/domain/usecase/GetPopularMovieUseCaseTest.kt
@@ -1,3 +1,5 @@
+package domain.usecase
+
 import edu.dyds.movies.domain.usecase.GetPopularMoviesUseCaseImpl
 import edu.dyds.movies.domain.entity.Movie
 import edu.dyds.movies.domain.repository.MoviesRepository
